+++ lib/src/tile_map.dart
@@ -8,6 +8,7 @@ class TileMap {
   List<Tileset> tilesets = new List<Tileset>();
   List<Layer> layers = new List<Layer>();
   List<ObjectGroup> objectGroups = [];
+  Map<String, dynamic> properties = {};
 
   /**
    * Retrieve a tile based on its GID
@@ -20,7 +21,9 @@ class TileMap {
    * A GID of 0 is always an "empty" tile
    */
   Tile getTileByGID(int gid) {
-    if (gid == 0) { return new Tile.emptyTile(); }
+    if (gid == 0) {
+      return new Tile.emptyTile();
+    }
     var ts = tilesets.lastWhere((tileset) => tileset.firstgid <= gid);
     return new Tile(gid - ts.firstgid, ts, false, false, false);
   }
@@ -28,8 +31,7 @@ class TileMap {
   // Returns a tileset based on its name
   Tileset getTileset(String name) {
     return tilesets.firstWhere((tileset) => tileset.name == name,
-      orElse: () => throw new ArgumentError('Tileset $name not found')
-    );
+        orElse: () => throw new ArgumentError('Tileset $name not found'));
   }
 
   /**
@@ -54,11 +56,15 @@ class TileMap {
   */
   Tile getTileByPhrase(String tilePhrase) {
     var split = tilePhrase.split('|');
-    if (split.length != 2) { throw new ArgumentError("$tilePhrase not in the format of 'TilesetName|LocalTileID"); }
+    if (split.length != 2) {
+      throw new ArgumentError(
+        '$tilePhrase not in the format of "TilesetName|LocalTileID"',
+      );
+    }
 
     var tilesetName = split.first;
-    var tileId = int.parse(split.last,
-      onError: (src) => throw new ArgumentError('Local tile ID $src is not an integer.') );
+    var tileId =
+        int.parse(split.last, onError: (src) => throw new ArgumentError('Local tile ID $src is not an integer.'));
 
     return getTileByLocalID(tilesetName, tileId);
   }