+++ test/layer_test.dart
@@ -5,14 +5,14 @@ import 'dart:async';
 import 'dart:io';
 
 main() {
-  var xmlRoot, xmlRoot_base64_gzip;
+  XmlElement xmlRoot, xmlRoot_base64_gzip;
 
   // Urgh. var xml = File.read(/* ... */); >:/
-  setUp( () {
-    var f1 = new File('./test/fixtures/test.tmx').readAsString().then((xml) {
+  setUp(() {
+    Future f1 = new File('./test/fixtures/test.tmx').readAsString().then((xml) {
       xmlRoot = parse(xml).rootElement;
     });
-    var f2 = new File('./test/fixtures/test_base64_gzip.tmx').readAsString().then((xml) {
+    Future f2 = new File('./test/fixtures/test_base64_gzip.tmx').readAsString().then((xml) {
       xmlRoot_base64_gzip = parse(xml).rootElement;
     });
 
@@ -21,8 +21,8 @@ main() {
 
   group('Layer.fromXML', () {
     test('supports gzip', () {
-      var layerNode = xmlRoot_base64_gzip.findAllElements('layer').first;
-      var layer = new Layer.fromXML(layerNode);
+      XmlNode layerNode = xmlRoot_base64_gzip.findAllElements('layer').first;
+      Layer layer = new Layer.fromXML(layerNode);
 
       expect(layer.tileMatrix[0], equals([1, 0, 0, 0, 0, 0, 0, 0, 0, 0]));
       expect(layer.tileMatrix[1], equals([0, 1, 0, 0, 0, 0, 0, 0, 0, 0]));
@@ -34,11 +34,10 @@ main() {
       expect(layer.tileMatrix[7], equals([0, 0, 0, 0, 0, 0, 0, 1, 0, 0]));
       expect(layer.tileMatrix[8], equals([0, 0, 0, 0, 0, 0, 0, 0, 1, 0]));
       expect(layer.tileMatrix[9], equals([0, 0, 0, 0, 0, 0, 0, 0, 0, 1]));
-
     });
     test('supports zlib', () {
-      var layerNode = xmlRoot.findAllElements('layer').first;
-      var layer = new Layer.fromXML(layerNode);
+      XmlNode layerNode = xmlRoot.findAllElements('layer').first;
+      Layer layer = Layer.fromXML(layerNode);
 
       expect(layer.tileMatrix[0], equals([1, 0, 0, 0, 0, 0, 0, 0, 0, 0]));
       expect(layer.tileMatrix[1], equals([0, 1, 0, 0, 0, 0, 0, 0, 0, 0]));
@@ -54,25 +53,30 @@ main() {
   });
 
   group('Layer.tiles', () {
-    var map, layer;
+    TileMap map;
+    Layer layer;
+
     setUp(() {
-      map = new TileMapParser().parse(xmlRoot.toString());
+      map = TileMapParser().parse(xmlRoot.toString());
       layer = map.layers.first;
     });
 
-
     test('is the expected size of 100', () {
-      expect(layer.tiles.length, equals(100));
+      expect(layer.tiles.length, equals(10));
+      layer.tiles.forEach((row) {
+        expect(row.length, equals(10));
+      });
     });
 
     test('calculates the x and y correctly for every tile', () {
-      var coords = [];
-      var expectedCoords = [];
-      layer.tiles.forEach((tile) => coords.add([tile.x, tile.y]));
+      List<List<int>> coords = [];
+      layer.tiles.forEach((row) => row.forEach((tile) => coords.add([tile.x, tile.y])));
+
       // Tileset is 32x32 in test.tmx, and the map is 10x10.
-      for(var y = 0; y < 10; y++) {
-        for(var x = 0; x < 10; x++) {
-          expectedCoords.add([x * 32, y * 32]);
+      List<List<int>> expectedCoords = [];
+      for (int x = 0; x < 10; x++) {
+        for (int y = 0; y < 10; y++) {
+          expectedCoords.add([x, y]);
         }
       }