+++ test/tile_test.dart
@@ -30,15 +30,15 @@ main() {
 
   test('Tile.properties is present', () {
     var tile = new Tile.emptyTile();
-    expect(tile.properties, new isInstanceOf<Map>());
+    expect(tile.properties, isA<Map>());
   });
 
   test('Tile.properties queries Tileset.tileProperties correctly', () {
-    var ts = new Tileset(1)..tileProperties[1] = { 'tile_property': 'tile_value' };
+    var ts = new Tileset(1)..tileProperties[1] = {'tile_property': 'tile_value'};
 
     var tile = new Tile(0, ts, false, false, false);
 
-    expect(tile.properties, equals({ 'tile_property': 'tile_value' }));
+    expect(tile.properties, equals({'tile_property': 'tile_value'}));
   });
 
   test('Tile.properties is an empty map if Tileset.tileProperties is empty for this tile', () {
@@ -46,5 +46,4 @@ main() {
     var tile = new Tile(2, ts, false, false, false);
     expect(tile.properties, equals({}));
   });
-  
 }